python_client/requirements.txt
paho-mqtt
cryptography
python-dotenv
